--- a/lib/cli-engine.js
+++ b/lib/cli-engine.js
@@ -17,12 +17,10 @@
 
 const fs = require("fs"),
     path = require("path"),
-    rules = require("./rules"),
-    eslint = require("./eslint"),
     defaultOptions = require("../conf/default-cli-options"),
+    Linter = require("./linter"),
     IgnoredPaths = require("./ignored-paths"),
     Config = require("./config"),
-    Plugins = require("./config/plugins"),
     fileEntryCache = require("file-entry-cache"),
     globUtil = require("./util/glob-util"),
     SourceCodeFixer = require("./util/source-code-fixer"),
@@ -143,11 +141,12 @@ function calculateStatsPerRun(results) {
  * @param {string} options.filename The filename from which the text was read.
  * @param {boolean} options.allowInlineConfig Flag indicating if inline comments
  *      should be allowed.
+ * @param {Linter} linter Linter context
  * @returns {Object} The result of the fix operation as returned from the
  *      SourceCodeFixer.
  * @private
  */
-function multipassFix(text, config, options) {
+function multipassFix(text, config, options, linter) {
     const MAX_PASSES = 10;
     let messages = [],
         fixedResult,
@@ -167,10 +166,10 @@ function multipassFix(text, config, options) {
         passNumber++;
 
         debug(`Linting code for ${options.filename} (pass ${passNumber})`);
-        messages = eslint.verify(text, config, options);
+        messages = linter.verify(text, config, options);
 
         debug(`Generating fixed text for ${options.filename} (pass ${passNumber})`);
-        fixedResult = SourceCodeFixer.applyFixes(eslint.getSourceCode(), messages);
+        fixedResult = SourceCodeFixer.applyFixes(linter.getSourceCode(), messages);
 
         // stop if there are any syntax errors.
         // 'fixedResult.output' is a empty string.
@@ -195,7 +194,7 @@ function multipassFix(text, config, options) {
      * the most up-to-date information.
      */
     if (fixedResult.fixed) {
-        fixedResult.messages = eslint.verify(text, config, options);
+        fixedResult.messages = linter.verify(text, config, options);
     }
 
 
@@ -214,13 +213,14 @@ function multipassFix(text, config, options) {
  * @param {string} filename An optional string representing the texts filename.
  * @param {boolean} fix Indicates if fixes should be processed.
  * @param {boolean} allowInlineConfig Allow/ignore comments that change config.
+ * @param {Linter} linter Linter context
  * @returns {LintResult} The results for linting on this text.
  * @private
  */
-function processText(text, configHelper, filename, fix, allowInlineConfig) {
+function processText(text, configHelper, filename, fix, allowInlineConfig, linter) {
 
     // clear all existing settings for a new file
-    eslint.reset();
+    linter.reset();
 
     let filePath,
         messages,
@@ -238,10 +238,10 @@ function processText(text, configHelper, filename, fix, allowInlineConfig) {
     const config = configHelper.getConfig(filePath);
 
     if (config.plugins) {
-        Plugins.loadAll(config.plugins);
+        configHelper.plugins.loadAll(config.plugins);
     }
 
-    const loadedPlugins = Plugins.getAll();
+    const loadedPlugins = configHelper.plugins.getAll();
 
     for (const plugin in loadedPlugins) {
         if (loadedPlugins[plugin].processors && Object.keys(loadedPlugins[plugin].processors).indexOf(fileExtension) >= 0) {
@@ -256,7 +256,7 @@ function processText(text, configHelper, filename, fix, allowInlineConfig) {
         const unprocessedMessages = [];
 
         parsedBlocks.forEach(block => {
-            unprocessedMessages.push(eslint.verify(block, config, {
+            unprocessedMessages.push(linter.verify(block, config, {
                 filename,
                 allowInlineConfig
             }));
@@ -272,10 +272,10 @@ function processText(text, configHelper, filename, fix, allowInlineConfig) {
             fixedResult = multipassFix(text, config, {
                 filename,
                 allowInlineConfig
-            });
+            }, linter);
             messages = fixedResult.messages;
         } else {
-            messages = eslint.verify(text, config, {
+            messages = linter.verify(text, config, {
                 filename,
                 allowInlineConfig
             });
@@ -310,13 +310,14 @@ function processText(text, configHelper, filename, fix, allowInlineConfig) {
  * @param {string} filename The filename of the file being checked.
  * @param {Object} configHelper The configuration options for ESLint.
  * @param {Object} options The CLIEngine options object.
+ * @param {Linter} linter Linter context
  * @returns {LintResult} The results for linting on this file.
  * @private
  */
-function processFile(filename, configHelper, options) {
+function processFile(filename, configHelper, options, linter) {
 
     const text = fs.readFileSync(path.resolve(filename), "utf8"),
-        result = processText(text, configHelper, filename, options.fix, options.allowInlineConfig);
+        result = processText(text, configHelper, filename, options.fix, options.allowInlineConfig, linter);
 
     return result;
 
@@ -471,6 +472,7 @@ class CLIEngine {
          * @type {Object}
          */
         this.options = options;
+        this.linter = new Linter();
 
         const cacheFile = getCacheFile(this.options.cacheLocation || this.options.cacheFile, this.options.cwd);
 
@@ -488,13 +490,15 @@ class CLIEngine {
 
             this.options.rulePaths.forEach(rulesdir => {
                 debug(`Loading rules from ${rulesdir}`);
-                rules.load(rulesdir, cwd);
+                this.linter.rules.load(rulesdir, cwd);
             });
         }
 
         Object.keys(this.options.rules || {}).forEach(name => {
-            validator.validateRuleOptions(name, this.options.rules[name], "CLI");
+            validator.validateRuleOptions(name, this.options.rules[name], "CLI", this.linter.rules);
         });
+
+        this.config = new Config(this.options, this.linter);
     }
 
     /**
@@ -542,8 +546,8 @@ class CLIEngine {
      * @param {Object} pluginobject Plugin configuration object.
      * @returns {void}
      */
-    addPlugin(name, pluginobject) { // eslint-disable-line class-methods-use-this
-        Plugins.define(name, pluginobject);
+    addPlugin(name, pluginobject) {
+        this.config.plugins.define(name, pluginobject);
     }
 
     /**
@@ -565,7 +569,7 @@ class CLIEngine {
         const results = [],
             options = this.options,
             fileCache = this._fileCache,
-            configHelper = new Config(options);
+            configHelper = this.config;
         let prevConfig; // the previous configuration used
 
         /**
@@ -602,9 +606,10 @@ class CLIEngine {
          * unsupported file extensions and any files that are already linted.
          * @param {string} filename The resolved filename of the file to be linted
          * @param {boolean} warnIgnored always warn when a file is ignored
+         * @param {Linter} linter Linter context
          * @returns {void}
          */
-        function executeOnFile(filename, warnIgnored) {
+        function executeOnFile(filename, warnIgnored, linter) {
             let hashOfConfig,
                 descriptor;
 
@@ -647,7 +652,7 @@ class CLIEngine {
 
             debug(`Processing ${filename}`);
 
-            const res = processFile(filename, configHelper, options);
+            const res = processFile(filename, configHelper, options, linter);
 
             if (options.cache) {
 
@@ -685,7 +690,7 @@ class CLIEngine {
         const fileList = globUtil.listFilesToProcess(patterns, options);
 
         fileList.forEach(fileInfo => {
-            executeOnFile(fileInfo.filename, fileInfo.ignored);
+            executeOnFile(fileInfo.filename, fileInfo.ignored, this.linter);
         });
 
         const stats = calculateStatsPerRun(results);
@@ -718,7 +723,7 @@ class CLIEngine {
 
         const results = [],
             options = this.options,
-            configHelper = new Config(options),
+            configHelper = this.config,
             ignoredPaths = new IgnoredPaths(options);
 
         // resolve filename based on options.cwd (for reporting, ignoredPaths also resolves)
@@ -731,7 +736,7 @@ class CLIEngine {
                 results.push(createIgnoreResult(filename, options.cwd));
             }
         } else {
-            results.push(processText(text, configHelper, filename, options.fix, options.allowInlineConfig));
+            results.push(processText(text, configHelper, filename, options.fix, options.allowInlineConfig, this.linter));
         }
 
         const stats = calculateStatsPerRun(results);
@@ -753,7 +758,7 @@ class CLIEngine {
      * @returns {Object} A configuration object for the file.
      */
     getConfigForFile(filePath) {
-        const configHelper = new Config(this.options);
+        const configHelper = this.config;
 
         return configHelper.getConfig(filePath);
     }