--- a/lib/config.js
+++ b/lib/config.js
@@ -43,10 +43,11 @@ function isObject(item) {
 /**
  * Load and parse a JSON config object from a file.
  * @param {string|Object} configToLoad the path to the JSON config file or the config object itself.
+ * @param {Config} configContext config instance object
  * @returns {Object} the parsed config object (empty object if there was a parse error)
  * @private
  */
-function loadConfig(configToLoad) {
+function loadConfig(configToLoad, configContext) {
     let config = {},
         filePath = "";
 
@@ -56,11 +57,11 @@ function loadConfig(configToLoad) {
             config = configToLoad;
 
             if (config.extends) {
-                config = ConfigFile.applyExtends(config, filePath);
+                config = ConfigFile.applyExtends(config, configContext, filePath);
             }
         } else {
             filePath = configToLoad;
-            config = ConfigFile.load(filePath);
+            config = ConfigFile.load(filePath, configContext);
         }
 
     }
@@ -70,10 +71,11 @@ function loadConfig(configToLoad) {
 
 /**
  * Get personal config object from ~/.eslintrc.
+ * @param {Config} configContext Plugin context for the config instance
  * @returns {Object} the personal config object (null if there is no personal config)
  * @private
  */
-function getPersonalConfig() {
+function getPersonalConfig(configContext) {
     let config;
 
     if (PERSONAL_CONFIG_DIR) {
@@ -81,7 +83,7 @@ function getPersonalConfig() {
 
         if (filename) {
             debug("Using personal config");
-            config = loadConfig(filename);
+            config = loadConfig(filename, configContext);
         }
     }
 
@@ -99,7 +101,7 @@ function hasRules(options) {
 
 /**
  * Get a local config object.
- * @param {Object} thisConfig A Config object.
+ * @param {Config} thisConfig A Config object.
  * @param {string} directory The directory to start looking in for a local config file.
  * @returns {Object} The local config object, or an empty object if there is no local config.
  */
@@ -127,7 +129,7 @@ function getLocalConfig(thisConfig, directory) {
         }
 
         debug(`Loading ${localConfigFile}`);
-        const localConfig = loadConfig(localConfigFile);
+        const localConfig = loadConfig(localConfigFile, thisConfig);
 
         // Don't consider a local config file found if the config is null
         if (!localConfig) {
@@ -152,7 +154,7 @@ function getLocalConfig(thisConfig, directory) {
          * - Otherwise, if no rules were manually passed in, throw and error.
          * - Note: This function is not called if useEslintrc is false.
          */
-        const personalConfig = getPersonalConfig();
+        const personalConfig = getPersonalConfig(thisConfig);
 
         if (personalConfig) {
             config = ConfigOps.merge(config, personalConfig);
@@ -186,9 +188,13 @@ class Config {
     /**
      * Config options
      * @param {Object} options Options to be passed in
+     * @param {Linter} linterContext Linter instance object
      */
-    constructor(options) {
+    constructor(options, linterContext) {
         options = options || {};
+
+        this.linterContext = linterContext;
+        this.plugins = new Plugins(linterContext.environments, linterContext.rules);
 
         this.ignore = options.ignore;
         this.ignorePath = options.ignorePath;
@@ -196,7 +202,7 @@ class Config {
         this.parser = options.parser;
         this.parserOptions = options.parserOptions || {};
 
-        this.baseConfig = options.baseConfig ? loadConfig(options.baseConfig) : { rules: {} };
+        this.baseConfig = options.baseConfig ? loadConfig(options.baseConfig, this) : { rules: {} };
 
         this.useEslintrc = (options.useEslintrc !== false);
 
@@ -219,16 +225,23 @@ class Config {
             return globals;
         }, {});
 
-        const useConfig = options.configFile;
-
         this.options = options;
 
-        if (useConfig) {
-            debug(`Using command line config ${useConfig}`);
-            if (isResolvable(useConfig) || isResolvable(`eslint-config-${useConfig}`) || useConfig.charAt(0) === "@") {
-                this.useSpecificConfig = loadConfig(useConfig);
+        this.loadConfigFile(options.configFile);
+    }
+
+    /**
+     * Loads the config from the configuration file
+     * @param {string} configFile - patch to the config file
+     * @returns {undefined}
+     */
+    loadConfigFile(configFile) {
+        if (configFile) {
+            debug(`Using command line config ${configFile}`);
+            if (isResolvable(configFile) || isResolvable(`eslint-config-${configFile}`) || configFile.charAt(0) === "@") {
+                this.useSpecificConfig = loadConfig(configFile, this);
             } else {
-                this.useSpecificConfig = loadConfig(path.resolve(this.options.cwd, useConfig));
+                this.useSpecificConfig = loadConfig(path.resolve(this.options.cwd, configFile), this);
             }
         }
     }
@@ -306,13 +319,13 @@ class Config {
         // Step 8: Merge in command line plugins
         if (this.options.plugins) {
             debug("Merging command line plugins");
-            Plugins.loadAll(this.options.plugins);
+            this.plugins.loadAll(this.options.plugins);
             config = ConfigOps.merge(config, { plugins: this.options.plugins });
         }
 
         // Step 9: Apply environments to the config if present
         if (config.env) {
-            config = ConfigOps.applyEnvironments(config);
+            config = ConfigOps.applyEnvironments(config, this.linterContext.environments);
         }
 
         this.cache[directory] = config;