--- a/lib/config/autoconfig.js
+++ b/lib/config/autoconfig.js
@@ -10,12 +10,13 @@
 //------------------------------------------------------------------------------
 
 const lodash = require("lodash"),
-    eslint = require("../eslint"),
+    Linter = require("../linter"),
     configRule = require("./config-rule"),
     ConfigOps = require("./config-ops"),
     recConfig = require("../../conf/eslint-recommended");
 
 const debug = require("debug")("eslint:autoconfig");
+const linter = new Linter();
 
 //------------------------------------------------------------------------------
 // Data
@@ -290,7 +291,7 @@ class Registry {
 
             ruleSets.forEach(ruleSet => {
                 const lintConfig = Object.assign({}, config, { rules: ruleSet });
-                const lintResults = eslint.verify(sourceCodes[filename], lintConfig);
+                const lintResults = linter.verify(sourceCodes[filename], lintConfig);
 
                 lintResults.forEach(result => {