--- a/lib/config/config-file.js
+++ b/lib/config/config-file.js
@@ -15,7 +15,6 @@ const fs = require("fs"),
     path = require("path"),
     ConfigOps = require("./config-ops"),
     validator = require("./config-validator"),
-    Plugins = require("./plugins"),
     pathUtil = require("../util/path-util"),
     ModuleResolver = require("../util/module-resolver"),
     pathIsInside = require("path-is-inside"),
@@ -386,6 +385,7 @@ function getEslintCoreConfigPath(name) {
 /**
  * Applies values from the "extends" field in a configuration file.
  * @param {Object} config The configuration information.
+ * @param {Config} configContext Plugin context for the config instance
  * @param {string} filePath The file path from which the configuration information
  *      was loaded.
  * @param {string} [relativeTo] The path to resolve relative to.
@@ -393,7 +393,7 @@ function getEslintCoreConfigPath(name) {
  *      loaded and merged.
  * @private
  */
-function applyExtends(config, filePath, relativeTo) {
+function applyExtends(config, configContext, filePath, relativeTo) {
     let configExtends = config.extends;
 
     // normalize into an array for easier handling
@@ -418,7 +418,7 @@ function applyExtends(config, filePath, relativeTo) {
                 );
             }
             debug(`Loading ${parentPath}`);
-            return ConfigOps.merge(load(parentPath, false, relativeTo), previousValue);
+            return ConfigOps.merge(load(parentPath, configContext, false, relativeTo), previousValue);
         } catch (e) {
 
             /*
@@ -520,12 +520,12 @@ function resolve(filePath, relativeTo) {
  * Loads a configuration file from the given file path.
  * @param {string} filePath The filename or package name to load the configuration
  *      information from.
+ * @param {Config} configContext Plugins context
  * @param {boolean} [applyEnvironments=false] Set to true to merge in environment settings.
  * @param {string} [relativeTo] The path to resolve relative to.
  * @returns {Object} The configuration information.
- * @private
  */
-function load(filePath, applyEnvironments, relativeTo) {
+function load(filePath, configContext, applyEnvironments, relativeTo) {
     const resolvedPath = resolve(filePath, relativeTo),
         dirname = path.dirname(resolvedPath.filePath),
         lookupPath = getLookupPath(dirname);
@@ -535,7 +535,7 @@ function load(filePath, applyEnvironments, relativeTo) {
 
         // ensure plugins are properly loaded first
         if (config.plugins) {
-            Plugins.loadAll(config.plugins);
+            configContext.plugins.loadAll(config.plugins);
         }
 
         // include full path of parser if present
@@ -548,20 +548,20 @@ function load(filePath, applyEnvironments, relativeTo) {
         }
 
         // validate the configuration before continuing
-        validator.validate(config, filePath);
+        validator.validate(config, filePath, configContext.linterContext.rules, configContext.linterContext.environments);
 
         /*
          * If an `extends` property is defined, it represents a configuration file to use as
          * a "parent". Load the referenced file and merge the configuration recursively.
          */
         if (config.extends) {
-            config = applyExtends(config, filePath, dirname);
+            config = applyExtends(config, configContext, filePath, dirname);
         }
 
         if (config.env && applyEnvironments) {
 
             // Merge in environment-specific globals and parserOptions.
-            config = ConfigOps.applyEnvironments(config);
+            config = ConfigOps.applyEnvironments(config, configContext.linterContext.environments);
         }
 
     }