--- a/lib/config/environments.js
+++ b/lib/config/environments.js
@@ -11,32 +11,30 @@
 const envs = require("../../conf/environments");
 
 //------------------------------------------------------------------------------
-// Private
-//------------------------------------------------------------------------------
-
-let environments = new Map();
-
-/**
- * Loads the default environments.
- * @returns {void}
- * @private
- */
-function load() {
-    Object.keys(envs).forEach(envName => {
-        environments.set(envName, envs[envName]);
-    });
-}
-
-// always load default environments upfront
-load();
-
-//------------------------------------------------------------------------------
 // Public Interface
 //------------------------------------------------------------------------------
 
-module.exports = {
+class Environments {
 
-    load,
+    /**
+     * create env context
+     */
+    constructor() {
+        this._environments = new Map();
+
+        this.load();
+    }
+
+    /**
+     * Loads the default environments.
+     * @returns {void}
+     * @private
+     */
+    load() {
+        Object.keys(envs).forEach(envName => {
+            this._environments.set(envName, envs[envName]);
+        });
+    }
 
     /**
      * Gets the environment with the given name.
@@ -44,8 +42,19 @@ module.exports = {
      * @returns {Object?} The environment object or null if not found.
      */
     get(name) {
-        return environments.get(name) || null;
-    },
+        return this._environments.get(name) || null;
+    }
+
+    /**
+     * Gets all the environment present
+     * @returns {Object} The environment object for each env name
+     */
+    getAll() {
+        return Array.from(this._environments).reduce((coll, env) => {
+            coll[env[0]] = env[1];
+            return coll;
+        }, {});
+    }
 
     /**
      * Defines an environment.
@@ -54,8 +63,8 @@ module.exports = {
      * @returns {void}
      */
     define(name, env) {
-        environments.set(name, env);
-    },
+        this._environments.set(name, env);
+    }
 
     /**
      * Imports all environments from a plugin.
@@ -69,14 +78,7 @@ module.exports = {
                 this.define(`${pluginName}/${envName}`, plugin.environments[envName]);
             });
         }
-    },
-
-    /**
-     * Resets all environments. Only use for tests!
-     * @returns {void}
-     */
-    testReset() {
-        environments = new Map();
-        load();
     }
-};
+}
+
+module.exports = Environments;