--- a/lib/load-rules.js
+++ b/lib/load-rules.js
@@ -12,6 +12,8 @@
 const fs = require("fs"),
     path = require("path");
 
+const rulesDirCache = {};
+
 //------------------------------------------------------------------------------
 // Public Interface
 //------------------------------------------------------------------------------
@@ -29,6 +31,11 @@ module.exports = function(rulesDir, cwd) {
         rulesDir = path.resolve(cwd, rulesDir);
     }
 
+    // cache will help performance as IO operation are expensive
+    if (rulesDirCache[rulesDir]) {
+        return rulesDirCache[rulesDir];
+    }
+
     const rules = Object.create(null);
 
     fs.readdirSync(rulesDir).forEach(file => {
@@ -37,5 +44,7 @@ module.exports = function(rulesDir, cwd) {
         }
         rules[file.slice(0, -3)] = path.join(rulesDir, file);
     });
+    rulesDirCache[rulesDir] = rules;
+
     return rules;
 };