--- a/lib/util/source-code-util.js
+++ b/lib/util/source-code-util.js
@@ -10,7 +10,6 @@
 //------------------------------------------------------------------------------
 
 const CLIEngine = require("../cli-engine"),
-    eslint = require("../eslint"),
     globUtil = require("./glob-util"),
     baseDefaultOptions = require("../../conf/default-cli-options");
 
@@ -38,7 +37,7 @@ function getSourceCodeOfFile(filename, options) {
 
         throw new Error(`(${filename}:${msg.line}:${msg.column}) ${msg.message}`);
     }
-    const sourceCode = eslint.getSourceCode();
+    const sourceCode = cli.linter.getSourceCode();
 
     return sourceCode;
 }