--- a/tests/lib/cli.js
+++ b/tests/lib/cli.js
@@ -19,8 +19,7 @@ const assert = require("chai").assert,
     leche = require("leche"),
     fs = require("fs"),
     os = require("os"),
-    sh = require("shelljs"),
-    rules = require("../../lib/rules");
+    sh = require("shelljs");
 
 const proxyquire = require("proxyquire").noCallThru().noPreserveCache();
 
@@ -92,7 +91,6 @@ describe("cli", () => {
 
     after(() => {
         sh.rm("-r", fixtureDir);
-        rules.testReset();
     });
 
     describe("execute()", () => {