### Linked issues

* ~https://github.com/eslint/eslint.github.io/issues/322~
* https://github.com/eslint/eslint/issues/8454

### Linter class

* Moved all the logic from `eslint` file into a `class` form to `Linter`.
* ~Now `eslint` file exposes itself by using linter as `new Linter()`.~
* Nothing breaks right now.
* Since `CLIEngine` consumes `Linter` now then it makes `CLIEngine` immutable across different instances.

### TODO

- [x] Make `rules` module as a `class` then consume it in `Linter`.
- [x] Create `Linter` class
- [x] Fix existing Unit test
- [x] Remove `eslint` file and make `CLIEngine` consume `Linter`. That way every instance of CLIEngine gets its own linter.

**If accepted:**
- [x] Unit test
- [x] Docs