@ilyavolodin 2017-04-15
I'm somewhat confused by this. The main reason why we can't do some of the things in web demo, is because multi-pass autofixing is done in CLIEngine, which is not browserifiable. Why do we need a separate `Linter` class to fix that?
Right now, it seems like you removed everything from eslint.js file and moved it into Linter.js. Do we still need eslint.js in that case?