@gyandeeps 2017-04-15
`eslint` module exists for backward compatibility.