@platinumazure 2017-04-15
@ilyavolodin It may not directly achieve the stated goal, but I think this is a step in the right direction to solve some of the mutable state problems @not-an-aardvark has called out. It's ultimately up to TSC, but for what it's worth, I'm :+1: to this change.