@not-an-aardvark 2017-04-16
I think this is a good idea, but can we separate the web engine functionality into a different PR? Making `Linter` an ES6 class and creating a `WebEngine` only seem loosely related to me, and I don't want them to block each other.