@gyandeeps 2017-04-21
I was thinking to make this as part of 4.0 release as well so that we dont impact anyone who uses mutability to its advantage if we release this as a minor. I bet nobody will get impacted but i am just being cautious.
@eslint/eslint-team thoughts on this PR?