@gyandeeps 2017-05-12
### Performance budget

#### Before

```
Loading:
  Load performance Run #1:  227.912401ms
  Load performance Run #2:  232.151311ms
  Load performance Run #3:  236.295281ms
  Load performance Run #4:  229.252669ms
  Load performance Run #5:  229.977971ms

  Load Performance median:  229.977971ms

Single File:
  CPU Speed is 2394 with multiplier 13000000
  Performance Run #1:  4897.325322ms
  Performance Run #2:  4911.44433ms
  Performance Run #3:  4906.176912ms
  Performance Run #4:  4959.677326ms
  Performance Run #5:  5004.119167ms

  Performance budget ok:  4911.44433ms (limit: 5430.242272347536ms)

Multi Files (0 files):
  CPU Speed is 2394 with multiplier 39000000
  Performance Run #1:  13646.079244ms
  Performance Run #2:  14079.363668ms
  Performance Run #3:  13754.637163ms
  Performance Run #4:  13092.094083ms
  Performance Run #5:  13259.704654ms

  Performance budget ok:  13646.079244ms (limit: 16290.726817042607ms)
```

#### After

```
Loading:
  Load performance Run #1:  230.841406ms
  Load performance Run #2:  229.987808ms
  Load performance Run #3:  234.446104ms
  Load performance Run #4:  230.948746ms
  Load performance Run #5:  227.209338ms

  Load Performance median:  230.841406ms

Single File:
  CPU Speed is 2394 with multiplier 13000000
  Performance Run #1:  4947.089321ms
  Performance Run #2:  4902.697088ms
  Performance Run #3:  4929.7710289999995ms
  Performance Run #4:  4953.978832ms
  Performance Run #5:  4912.9244419999995ms

  Performance budget ok:  4929.7710289999995ms (limit: 5430.242272347536ms)

Multi Files (0 files):
  CPU Speed is 2394 with multiplier 39000000
  Performance Run #1:  16898.826388ms
  Performance Run #2:  13128.130804ms
  Performance Run #3:  13473.125132ms
  Performance Run #4:  13407.889025ms
  Performance Run #5:  13018.129979ms

  Performance budget ok:  13407.889025ms (limit: 16290.726817042607ms)
```