@not-an-aardvark 2017-05-16
@gyandeeps I'm planning to review this, but from the checkboxes in the original PR I see that it doesn't have docs and tests yet. Just to make sure, are those checkboxes still accurate?