@gyandeeps 2017-05-16
@not-an-aardvark checkboxes are correct. I have not added test and docs yet. Most of the code is already tested, the test i will add will only verify the mutability part.
You can go ahead to review this PR.