@not-an-aardvark 2017-05-16
Ok, I'll review it now.

Also, I just want to make sure you're aware of the potential compatibility issue from https://github.com/eslint/eslint/issues/6237#issuecomment-301371444.

One possible workaround would be to add a static `define` method to `Plugins` that keeps a global namespace as a fallback while still using individual instances of `Plugins` for most purposes.